New constraints/validations for Dependency resource Added 'DeleteMany' API

+++ app/Http/Controllers/DependencyController.php
@@ -16,7 +16,9 @@ class DependencyController extends Controller
     {
         $dependencies = Dependency::orderBy('name', 'asc')->get();
 
-        return $dependencies;
+        return response() -> json( [
+            'data' => $dependencies,
+        ], 200);
     }
 
     /**
@@ -35,17 +37,20 @@ class DependencyController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store( Request $request )
     {
         $validated = $request->validate([
-            'name' => ['bail', 'required', 'max:255'],
+            'name' => [ 'required' ],
         ]);
 
         $dependency = new Dependency();
         $dependency->name = $request->name;
         $dependency->save();
 
-        return $dependency;
+        return response() -> json( [
+            'message' => 'Dependency created succesfully.',
+            'data' => $dependency,
+        ], 200);
     }
 
     /**
@@ -54,11 +59,13 @@ class DependencyController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show( $id )
     {
-        $dependency = Dependency::where('_id', $id)->get();
+        $dependency = Dependency::where( '_id', $id )->get();
 
-        return $dependency;
+        return response() -> json( [
+            'data' => $dependency
+        ], 200);
     }
 
     /**
@@ -79,13 +86,26 @@ class DependencyController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update( Request $request, $id )
     {
-        $dependency = Dependency::find($id);
-        
+        $dependency = Dependency::find( $id );
+
+        if ( $dependency == null )
+        {
+            return response()->json( $this->handleErrors( 'notfound' ), 404 );
+        }
+
+        $validated = $request->validate([
+            'name' => [ 'required' ],
+        ]);
+
         $dependency->name = $request->name;
         $dependency->save();
-        return $dependency;
+
+        return response() -> json( [
+            'message' => 'Dependency updated succesfully.',
+            'data' => $dependency,
+        ], 200);
     }
 
     /**
@@ -94,12 +114,55 @@ class DependencyController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy( $id )
     {
-        $dependency = Dependency::find($id);
+        $dependency = Dependency::find( $id );
+
+        if ( $dependency == null )
+        {
+            return response()->json( $this->handleErrors( 'notfound' ), 404 );
+        }
 
         $dependency->delete();
 
-        return $dependency;
+        return response()->json( [ 'message' => 'Dependency deleted succesfully.' ], 200);
+    }
+
+    public function destroyMany( Request $request )
+	{
+        $validated = $request->validate([
+            'data' => [ 'required' ]
+        ]);
+
+        $data = $request->data;
+        
+		foreach ($data as $id) 
+		{
+            if (array_key_exists('_id', $id))
+            {
+                Dependency::where('_id', $id['_id'])->delete();
+            }
+        }
+        
+		return response()->json( [ 'message' => 'Dependencies deleted succesfully.' ], 200);
+	}
+
+    public function handleErrors( $error )
+    {
+        switch ( $error )
+        {
+            case 'notfound':
+            {
+                return [
+                    'message' => 'The given data was invalid.',
+                    'errors' =>
+                        [
+                            'id' => 'There isn\'t a Dependency associated with that id.',
+                        ]
+                ];
+
+                break;
+            }
+        }
     }
 }
